Hackaton15/aGimenez/appTipos/views.py
from rest_framework import generics
from .models import Tipo
from .serializers import TipoSerializer


class TipoListCreateAPIView(generics.ListCreateAPIView):
    queryset = Tipo.objects.all()
    serializer_class = TipoSerializer


class TipoRetrieveUpdateDestroyAPIView(generics.RetrieveUpdateDestroyAPIView):
    queryset = Tipo.objects.all()
    serializer_class = TipoSerializer
