Hackaton15/aGimenez/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-environ==0.10.0
django-filter==23.2
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
